model-training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
hydra-core>=1.3.0
timm>=0.9.0
dvc>=3.0.0
dvc[s3]>=3.0.0
scikit-learn>=1.0.0 